Update memory management helpers to use 32 bit addresses

+++ tile/kernel/memory.h
@@ -53,8 +53,8 @@ enum memory_map_block_flags {
   struct memory_map_block represents the bounds of a memory block.
 */
 struct memory_map_block {
-  uint64_t begin;
-  uint64_t size;
+  uint32_t begin;
+  uint32_t size;
   int flags;
 };
 
@@ -70,7 +70,7 @@ struct memory_map_group {
   struct memory_map represents the memory map of a machine.
 */
 struct memory_map {
-  uint64_t limit;
+  uint32_t limit;
   struct memory_map_group* memory;
   struct memory_map_group* reserved;
 };
@@ -101,8 +101,8 @@ struct memory_manager {
   uint32_t bss_end;
 };
 
-uint64_t virt_to_phys(uint32_t x);
-uint32_t phys_to_virt(uint64_t x);
+uint32_t virt_to_phys(uint32_t x);
+uint32_t phys_to_virt(uint32_t x);
 
 void init_memory_map();
 void init_memory_manager(void* pgd, void* text_begin, void* text_end, void* data_begin, void* data_end, void* bss_begin, void* bss_end);
@@ -111,9 +111,9 @@ void update_memory_map();
 
 void memory_map_mask_block(struct memory_map_block* block, int flag, int mask);
 void memory_map_merge_blocks(struct memory_map_group* group, int begin, int end);
-void memory_map_insert_block(struct memory_map_group* group, int pos, uint64_t begin, uint64_t size);
-void memory_map_add_block(struct memory_map_group* group, uint64_t begin, uint64_t size);
-int memory_map_split_block(struct memory_map_group* group, uint64_t begin);
+void memory_map_insert_block(struct memory_map_group* group, int pos, uint32_t begin, uint32_t size);
+void memory_map_add_block(struct memory_map_group* group, uint32_t begin, uint32_t size);
+int memory_map_split_block(struct memory_map_group* group, uint32_t begin);
 
 size_t bitmap_index(const struct memory_bitmap* bitmap, uint32_t addr);
 size_t bitmap_index_index(const struct memory_bitmap* bitmap, uint32_t addr);
